Rename sudoku.cpp to main.cpp
